+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/entities/mobs/BeeQueenEntity.java
@@ -72,6 +72,7 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.GameRules;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.LevelReader;
@@ -499,7 +500,10 @@ public class BeeQueenEntity extends Animal implements NeutralMob {
         }
 
         int beeCooldown = this.getBeeSpawnCooldown();
-        if (beeCooldown <= 0 && !this.isImmobile()) {
+        if (beeCooldown <= 0 &&
+            !this.isImmobile() &&
+            this.level().getGameRules().getBoolean(GameRules.RULE_DOMOBSPAWNING))
+        {
             this.setBeeSpawnCooldown(this.random.nextInt(50) + 75);
 
             // Grab a nearby air materialposition a bit away